阿里云+亚马逊oss文件上传

+++ whiz_service/whiz_service_file/src/main/java/com/damon/file/pojo/FileInfo.java
@@ -35,7 +35,7 @@ public class FileInfo implements Serializable {
     /**
      * 文件大小
      */
-    private long size;
+    private Integer size;
     /**
      * 冗余字段
      */
